+++ packages/fighting-design/divider/index.ts
@@ -4,6 +4,7 @@ import { install } from '../_utils'
 
 export const FDivider = install(Divider)
 
+/** divider 组件实例类型 */
 export type DividerInstance = InstanceType<typeof Divider>
 
 export * from './src/interface.d'